PreKernel: Move Prekernel subdirectory outside Kernel directory
I want the PreKernel environment to eventually stand on its own. It will
always be related to the Kernel, but the plan is to make it a completely
standalone environment - in compile time and in runtime.
