+++ Kernel/CMakeLists.txt
@@ -496,5 +496,3 @@ install(FILES "${CMAKE_CURRENT_BINARY_DIR}/kernel.map" DESTINATION res)
 
 serenity_install_headers(Kernel)
 serenity_install_sources(Kernel)
-
-add_subdirectory(Prekernel)
