+++ Kernel/Memory/PageDirectory.cpp
@@ -8,10 +8,10 @@
 #include <AK/Singleton.h>
 #include <Kernel/Memory/MemoryManager.h>
 #include <Kernel/Memory/PageDirectory.h>
-#include <Kernel/Prekernel/Prekernel.h>
 #include <Kernel/Process.h>
 #include <Kernel/Random.h>
 #include <Kernel/Sections.h>
+#include <PreKernel/Prekernel.h>
 
 extern u8 end_of_kernel_image[];
