+++ PreKernel/StdLibExtras.h
